Fixing warning for implicit double to size_t conversion in read_simulator.h.
git-svn-id: http://svn.seqan.de/seqan/trunk@7203 e6417c60-b987-48fd-844e-b20f0fcc1017

+++ seqan/projects/library/apps/read_simulator/read_simulator.h
@@ -622,7 +622,7 @@ int buildReadSimulationInstruction(
             return 1;
         }
         // Pick a start and end position.
-        inst.beginPos = pickRandomNumber(rng, PDF<Uniform<double> >(0, length(haplotype[inst.contigId]) - readLength - 1));
+        inst.beginPos = pickRandomNumber(rng, PDF<Uniform<size_t> >(0, length(haplotype[inst.contigId]) - readLength - 1));
         inst.endPos = inst.beginPos + readLength;
         // Simulate the read with these parameters.
         buildSimulationInstructions(inst, rng, readLength, haplotype[inst.contigId], parameters, options);
